Rework release workflow (#361)
* fix typo

* add release_type input choice

* adjust script accordingly

* add if check

* still need if statement

* fix args for script

* conditional build for docker

* add a generic task issue template

* update cargo

* bump runtime versions

+++ scripts/generate-release-notes.sh
@@ -2,16 +2,16 @@
 set -eo pipefail
 
 err_report() {
-    echo "Error on line $1"
+  echo "Error on line $1"
 }
 
 trap 'err_report $LINENO' ERR
 
 function usage() {
-    echo "Usage: $0 litentry|litmus path-to-output diff-tag"
+  echo "Usage: $0 litentry|litmus path-to-output diff-tag release-type"
 }
 
-[ $# -ne 3 ] && (usage; exit 1)
+[ $# -ne 3 ] && [ $# -ne 4 ] && (usage; exit 1)
 
 ROOTDIR=$(git rev-parse --show-toplevel)
 cd "$ROOTDIR"
@@ -19,31 +19,34 @@ cd "$ROOTDIR"
 CHAIN=$1
 REPO=https://github.com/litentry/litentry-parachain
 
-# base image used to build the node binary
-NODE_BUILD_BASE_IMAGE=$(grep FROM docker/Dockerfile | head -n1 | sed 's/^FROM //;s/ as.*//')
+if [ "$3" != "runtime" ]; then
+  # base image used to build the node binary
+  NODE_BUILD_BASE_IMAGE=$(grep FROM docker/Dockerfile | head -n1 | sed 's/^FROM //;s/ as.*//')
 
-# somehow `docker inspect` doesn't pull our litentry-parachain image sometimes
-docker pull "$NODE_BUILD_BASE_IMAGE"
-docker pull "litentry/litentry-parachain:$RELEASE_TAG"
+  # somehow `docker inspect` doesn't pull our litentry-parachain image sometimes
+  docker pull "$NODE_BUILD_BASE_IMAGE"
+  docker pull "litentry/litentry-parachain:$RELEASE_TAG"
 
-NODE_VERSION=$(grep version node/Cargo.toml | head -n1 | sed "s/'$//;s/.*'//")
-NODE_BIN=litentry-collator
-NODE_SHA1SUM=$(shasum litentry-collator/"$NODE_BIN" | awk '{print $1}')
-if [ -f rust-toolchain.toml ]; then
-  NODE_RUSTC_VERSION=$(rustc --version)
-else
-  NODE_RUSTC_VERSION=$(docker run --rm "$NODE_BUILD_BASE_IMAGE" rustup default nightly 2>&1 | grep " installed" | sed 's/.*installed - //')
+  NODE_VERSION=$(grep version node/Cargo.toml | head -n1 | sed "s/'$//;s/.*'//")
+  NODE_BIN=litentry-collator
+  NODE_SHA1SUM=$(shasum litentry-collator/"$NODE_BIN" | awk '{print $1}')
+  if [ -f rust-toolchain.toml ]; then
+    NODE_RUSTC_VERSION=$(rustc --version)
+  else
+    NODE_RUSTC_VERSION=$(docker run --rm "$NODE_BUILD_BASE_IMAGE" rustup default nightly 2>&1 | grep " installed" | sed 's/.*installed - //')
+  fi
 fi
 
-SRTOOL_DIGEST_FILE=$CHAIN-parachain-runtime/$CHAIN-parachain-srtool-digest.json
-
-RUNTIME_VERSION=$(grep spec_version runtime/$CHAIN/src/lib.rs | sed 's/.*version: //;s/,//')
-RUNTIME_COMPRESSED_SIZE=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.size | sed 's/"//g')
-RUNTIME_RUSTC_VERSION=$(cat "$SRTOOL_DIGEST_FILE" | jq .rustc | sed 's/"//g')
-RUNTIME_COMPRESSED_SHA256=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.sha256 | sed 's/"//g')
-RUNTIME_COMPRESSED_BLAKE2=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.blake2_256 | sed 's/"//g')
-RUNTIME_COMPRESSED_SET_CODE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.proposal_hash | sed 's/"//g')
-RUNTIME_COMPRESSED_AUTHORIZE_UPGRADE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.parachain_authorize_upgrade_hash | sed 's/"//g')
+if [ "$3" != "client" ]; then
+  SRTOOL_DIGEST_FILE=$CHAIN-parachain-runtime/$CHAIN-parachain-srtool-digest.json
+  RUNTIME_VERSION=$(grep spec_version runtime/$CHAIN/src/lib.rs | sed 's/.*version: //;s/,//')
+  RUNTIME_COMPRESSED_SIZE=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.size | sed 's/"//g')
+  RUNTIME_RUSTC_VERSION=$(cat "$SRTOOL_DIGEST_FILE" | jq .rustc | sed 's/"//g')
+  RUNTIME_COMPRESSED_SHA256=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.sha256 | sed 's/"//g')
+  RUNTIME_COMPRESSED_BLAKE2=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.blake2_256 | sed 's/"//g')
+  RUNTIME_COMPRESSED_SET_CODE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.proposal_hash | sed 's/"//g')
+  RUNTIME_COMPRESSED_AUTHORIZE_UPGRADE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.parachain_authorize_upgrade_hash | sed 's/"//g')
+fi
 
 SUBSTRATE_DEP=$(grep sp-core node/Cargo.toml | sed 's/.*branch = "//;s/".*//')
 CUMULUS_DEP=$(grep cumulus-client-cli node/Cargo.toml | sed 's/.*branch = "//;s/".*//')
@@ -51,10 +54,13 @@ POLKADOT_DEP=$(grep polkadot-cli node/Cargo.toml | sed 's/.*branch = "//;s/".*//
 
 TAB="$(printf '\t')"
 
+echo > "$2"
+
 # use <CODE> to decorate around the stuff and then replace it with `
 # so that it's not executed as commands inside heredoc
-cat << EOF > "$2"
 
+if [ "$3" != "runtime" ]; then
+  cat << EOF >> "$2"
 ## Client
 
 <CODEBLOCK>
@@ -65,6 +71,11 @@ sha1sum                      : $NODE_SHA1SUM
 docker image                 : litentry/litentry-parachain:$RELEASE_TAG
 <CODEBLOCK>
 
+EOF
+fi
+
+if [ "$3" != "client" ]; then
+  cat << EOF >> "$2"
 ## Runtime
 
 <CODEBLOCK>
@@ -77,6 +88,10 @@ proposal (setCode)           : $RUNTIME_COMPRESSED_SET_CODE_HASH
 proposal (authorizeUpgrade)  : $RUNTIME_COMPRESSED_AUTHORIZE_UPGRADE_HASH
 <CODEBLOCK>
 
+EOF
+fi
+
+cat << EOF >> "$2"
 ## Dependencies
 
 <CODEBLOCK>
@@ -88,6 +103,11 @@ Cumulus                      : $CUMULUS_DEP
 EOF
 
 if [ "$GENESIS_RELEASE" = "true" ]; then
+  if [ "$3" = "runtime" ]; then
+    echo "genesis release requires to build client"
+    exit 1
+  fi
+
   GENESIS_STATE_HASH=$(shasum litentry-collator/$CHAIN-genesis-state | awk '{print $1}')
   GENESIS_WASM_HASH=$(shasum litentry-collator/$CHAIN-genesis-wasm | awk '{print $1}')
 
@@ -120,7 +140,7 @@ sed -i.bak 's/<CODEBLOCK>/```/g' "$2"
 rm -f "$2.bak"
 
 # if we have a diff-tag, list the changes inbetween
-DIFF_TAG="$3"
+DIFF_TAG="$4"
 
 if [ -z "$DIFF_TAG" ]; then
   echo "Nothing to compare"
